backend/requirements.txt
Flask
Flask-Cors
requests
beautifulsoup4
google-generativeai
certifi
gunicorn